+++ src/renderer/src/App.tsx
@@ -5,18 +5,21 @@ import { WorkspaceProvider } from './store/workspace'
 import { DiagnosticsProvider } from './store/diagnostics'
 import { SettingsProvider } from './store/settings'
 import { ConsoleProvider } from './store/console'
+import { SyncProvider } from './store/sync'
 
 function App(): JSX.Element {
   return (
     <PromptProvider>
       <SettingsProvider>
         <WorkspaceProvider>
-          <DiagnosticsProvider>
-            <ConsoleProvider>
-              <AppShell />
-              <UpdateNotifier />
-            </ConsoleProvider>
-          </DiagnosticsProvider>
+          <SyncProvider>
+            <DiagnosticsProvider>
+              <ConsoleProvider>
+                <AppShell />
+                <UpdateNotifier />
+              </ConsoleProvider>
+            </DiagnosticsProvider>
+          </SyncProvider>
         </WorkspaceProvider>
       </SettingsProvider>
     </PromptProvider>